Litigants App
- Moves the Litigant and LitigantRole into the new litigants app.
- Updates the references (including lazy references in migrations).
- Moves the serializers, views and urls.
- All unit tests passing.
