+++ conftest.py
@@ -5,13 +5,14 @@ from rest_framework.test import APIClient
 
 from core.models import User
 from events.tests.factories import TimelineFactory
-from poc.models import Case, Litigant, LitigantRole
+from litigants.models import Litigant, LitigantRole
+from litigants.tests.factories import LitigantFactory
+from poc.models import Case
 from poc.tests.factories import (
     CaseFactory,
     CaseLitigantFactory,
     ChatMessageFactory,
     ChatThreadFactory,
-    LitigantFactory,
     ParsedEmailAttachmentFactory,
     ParsedEmailFactory,
     UploadedFileFactory,
