+++ litigants/tests/api/views/test_create_litigants_api.py
@@ -2,7 +2,7 @@ from django.urls import reverse
 
 
 def _get_api_url():
-    return reverse("poc:litigants")
+    return reverse("litigants:list_create")
 
 
 def test_with_anonymous_user(api_client):
