+++ litigants/tests/factories.py
@@ -1,0 +1,6 @@
+import factory
+
+
+class LitigantFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = "litigants.Litigant"
