+++ poc/api/views.py
@@ -2,7 +2,6 @@ from django.db.models import Q
 from django.shortcuts import get_object_or_404
 from rest_framework.generics import (
     ListCreateAPIView,
-    RetrieveAPIView,
     RetrieveUpdateAPIView,
     RetrieveUpdateDestroyAPIView,
 )
@@ -16,11 +15,10 @@ from poc.api.serializers import (
     CaseSerializer,
     ChatMessageSerializer,
     ChatThreadSerializer,
-    LitigantSerializer,
     UploadedFileSerializer,
 )
 from poc.langchain.chat_agent import send_message
-from poc.models import Case, ChatMessage, ChatThread, Litigant, UploadedFile
+from poc.models import Case, ChatMessage, ChatThread, UploadedFile
 
 __all__ = [
     "ListCreateCaseAPI",
@@ -29,8 +27,6 @@ __all__ = [
     "RetrieveUpdateDestroyUploadedFileAPI",
     "ListCreateThreadAPI",
     "ListCreateMessageAPI",
-    "ListCreateLitigantAPI",
-    "RetrieveLitigantAPI",
 ]
 
 
@@ -222,35 +218,3 @@ class ListCreateMessageAPI(APIView):
         )
         op_serializer = ChatMessageSerializer(messages, many=True)
         return Response(op_serializer.data, status=201)
-
-
-class ListCreateLitigantAPI(ListCreateAPIView):
-    permission_classes = [IsAuthenticated]
-    serializer_class = LitigantSerializer
-
-    def get_queryset(self):
-        # check for query param 'search'
-        search = self.request.query_params.get("search")
-        if search:
-            if len(search.strip()) > 2:
-                # search in name, bio, email and phone fields
-                return Litigant.objects.filter(
-                    Q(name__icontains=search)
-                    | Q(bio__icontains=search)
-                    | Q(email__icontains=search)
-                    | Q(phone__icontains=search)
-                ).order_by("-id")
-            else:
-                return Litigant.objects.none()
-
-        return Litigant.objects.all().order_by("-id")
-
-
-class RetrieveLitigantAPI(RetrieveAPIView):
-    permission_classes = [IsAuthenticated]
-    serializer_class = LitigantSerializer
-    lookup_field = "id"
-    lookup_url_kwarg = "id"
-
-    def get_queryset(self):
-        return Litigant.objects.all()
